requirements.txt
anyio==3.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.1
filelock==3.12.0
fsspec==2023.5.0
h11==0.14.0
huggingface-hub==0.14.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pydantic==1.10.7
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
sniffio==1.3.0
starlette==0.26.1
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.29.0
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
